# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ALTIL Tenant, Governed AI and Communications API
  version: 1.1.0
  summary: Introsoft ALTIL multi-tenant AI SaaS platform
  description: >-
    Tenant onboarding, organization hierarchy, scoped API keys, license and quota
    enforcement, invoicing, governed AI inference and encrypted multi-channel communication.
    AI calls fail closed when tenant, application, key, billing or policy checks cannot pass.
servers:
  - url: https://api.introsoft.com
    description: Production API gateway
  - url: http://localhost:3005
    description: Local development
security:
  - ALTILKey: []
  - BearerSession: []
...
paths:
...
  /api/v1/public/registrations:
    post:
      tags: [Public SaaS]
      summary: Register a new individual or company tenant
      description: Creates a 14-day trial workspace, tenant owner, first application and scoped API key. The full key is returned once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, accountType, planId, applicationName, acceptedTerms]
              properties:
                name: {type: string, minLength: 2, maxLength: 160}
                email: {type: string, format: email}
                password: {type: string, minLength: 12, maxLength: 128, writeOnly: true}
                accountType: {type: string, enum: [individual, company]}
                planId: {type: string}
                applicationName: {type: string}
                country: {type: string}
                acceptedTerms: {type: boolean, const: true}
      responses:
        '201': {description: Tenant, owner, application, trial and one-time key provisioned}
        '400': {description: Invalid signup data or unpublished package}
        '409': {description: Owner email already exists}
        '429': {description: Signup rate limit}
        '503': {description: Durable database or encryption prerequisites unavailable}
...
components:
  securitySchemes:
    ALTILKey: {type: apiKey, in: header, name: x-api-key}
    BearerSession: {type: http, scheme: bearer, bearerFormat: ALTIL session}
    DeviceSecret: {type: apiKey, in: header, name: x-device-secret}
...